refactor(football): extract LineRenderer and build_football_field to separate modules
Split field_renderer.py (251→89 lines) by extracting LineRenderer to line_renderer.py,
and split player_renderer.py (195→148 lines) by extracting build_football_field to
build_field.py. DRY'd penalty arc drawing with _draw_single_penalty_arc helper.
